Remaining local-central items in Middleware. Added superclass DataManager.
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBPMiddleware@174 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/test/java/org/generationcp/middleware/manager/test/TestGermplasmAddUpdateFunctions.java
@@ -263,7 +263,8 @@ public class TestGermplasmAddUpdateFunctions
 		g.setLocationId(new Integer(9000));
 		g.setMethodId(new Integer(1));
 		g.setMgid(new Integer(1));
-		g.setUserId(new Integer(527));
+		//g.setUserId(new Integer(527));
+		g.setUserId(new Integer(1));
 		g.setReferenceId(new Integer(1));
 		
 		Name n = new Name();
